requirements.txt
git+http://root:NewPassword123%21@vectasurge.com:6080/review/common_util@0.1.0
asgiref==3.9.1
async-timeout==5.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.6
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.1
django-health-check==3.20.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
Markdown==3.9
mysqlclient==2.2.7
packaging==25.0
pika==1.3.2
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.2.1
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uWSGI==2.0.30
